Add dependency injection for GenericRole feature
It is expected that many users will want to integrate the GenericRole
feature with an own concept of 'groups' in their software. This
dependency injection allows these users to add some glue code which is
then made responsible for checking whether or not a user is part of some
Generic Role.

Closes #29.